refactor: DataFormat, TimeFormat 컴포넌트 제거 후 로직 utils로 분리
